Spring/eatgo/eatgo-api/src/main/java/me/choi/eatgo/application/RestaurantService.java
package me.choi.eatgo.application;

import me.choi.eatgo.domain.*;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.web.bind.annotation.GetMapping;

import java.util.List;

/**
 * Project : eatgo
 *
 * @author : jwdeveloper
 * @comment :
 * Time : 8:44 오후
 */
@Service
public class RestaurantService {

    @Autowired
    RestaurantRepository restaurantRepository;

    @Autowired
    MenuItemRepository menuItemRepository;
    

    public RestaurantService(RestaurantRepository restaurantRepository, MenuItemRepositoryImpl menuItemRepository) {
        this.restaurantRepository = restaurantRepository;
        this.menuItemRepository = menuItemRepository;
    }


    public Restaurant getRestaurant(Long id) {

        Restaurant restaurant = restaurantRepository.findById(id);

        List<MenuItem> menuItems = menuItemRepository.findAllByRepositoryId(id);
        restaurant.setMenuItem(menuItems);

        return restaurant;

    }

    public List<Restaurant> getRestaurants() {

        List<Restaurant> restaurants = restaurantRepository.findAll();

        return restaurants;
    }
}
